Conflicts:
	system/installer/updates/ud_260.php

+++ system/expressionengine/libraries/Filemanager.php
@@ -6,8 +6,8 @@
  * @package		ExpressionEngine
  * @author		EllisLab Dev Team
  * @copyright	Copyright (c) 2003 - 2012, EllisLab, Inc.
- * @license		http://expressionengine.com/user_guide/license.html
- * @link		http://expressionengine.com
+ * @license		http://ellislab.com/expressionengine/user-guide/license.html
+ * @link		http://ellislab.com
  * @since		Version 2.0
  * @filesource
  */
@@ -21,7 +21,7 @@
  * @subpackage	Core
  * @category	Filemanager
  * @author		EllisLab Dev Team
- * @link		http://expressionengine.com
+ * @link		http://ellislab.com
  */
 
 class Filemanager {
@@ -32,7 +32,6 @@ class Filemanager {
 	public $upload_errors		= FALSE;
 	public $upload_data			= NULL;
 	public $upload_warnings		= FALSE;
-	public $ignore_dupes		= TRUE;
 
 	private $EE;
 	
@@ -123,7 +122,7 @@ class Filemanager {
 		$ext = '.'.$ext;
 		
 		// Figure out a unique filename
-		if ($parameters['ignore_dupes'] === FALSE OR $this->ignore_dupes == FALSE)
+		if ($parameters['ignore_dupes'] === FALSE)
 		{
 			$basename = $filename;
 			
@@ -893,8 +892,8 @@ class Filemanager {
 
 		// Mask the URL if we're coming from the CP
 		$sync_files_url = (REQ == "CP") ?
-			$this->EE->cp->masked_url('http://expressionengine.com/user_guide/cp/content/files/sync_files.html') :
-			'http://expressionengine.com/user_guide/cp/content/files/sync_files.html';
+			$this->EE->cp->masked_url('http://ellislab.com/expressionengine/user-guide/cp/content/files/sync_files.html') :
+			'http://ellislab.com/expressionengine/user-guide/cp/content/files/sync_files.html';
 
 		return array(
 			'rows'			=> $this->_browser_get_files($dir, $file_params),
@@ -2343,6 +2342,13 @@ class Filemanager {
 		
 		// Delete the existing file's raw files, but leave the database record
 		$this->EE->file_model->delete_raw_file($file_name, $directory_id);
+
+		// It is possible the file exists but is NOT in the DB yet
+		if (empty($existing_file))
+		{
+			$new_file->modified_by_member_id = $this->EE->session->userdata('member_id');
+			return $new_file;
+		}
 		
 		// Delete the new file's database record, but leave the files
 		$this->EE->file_model->delete_files($new_file->file_id, FALSE);
@@ -2355,6 +2361,7 @@ class Filemanager {
 			'modified_date'			=> $new_file->modified_date,
 			'modified_by_member_id'	=> $this->EE->session->userdata('member_id')
 		));
+		
 		$existing_file->file_size				= $new_file->file_size;
 		$existing_file->file_hw_original		= $new_file->file_hw_original;
 		$existing_file->modified_date			= $new_file->modified_date;
